fix after review

+++ src/Converter.java
@@ -1,10 +1,10 @@
 public class Converter {
 
-    public int distanceCalculation(int steps){
-        return steps * 75/100000;
+    public double distanceCalculation(int steps) {
+        return steps * 75 / 100000;
     }
 
-    public int caloriesCalculation(int steps){
-        return steps * 50/1000;
+    public double caloriesCalculation(int steps) {
+        return steps * 50 / 1000;
     }
 }
